Safari-609.1.11.1/LayoutTests/http/tests/security/sandboxed-iframe-origin-add-expected.txt

CONSOLE MESSAGE: line 6: SecurityError: Sandbox access violation: Blocked a frame at "http://127.0.0.1:8000" from accessing a cross-origin frame.  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
Test that adding allow-origin after creating an iframe doesn't modify the origin of an existing document, but it doesn modify the origin of the next document.

PASS
